AMLS-4079 Split DataCacheConnector.scala into its component parts
