+++ app/services/cache/AmlsMongoCacheClient.scala
@@ -14,124 +14,22 @@
  * limitations under the License.
  */
 
-package connectors
+package services.cache
 
-import config.{AmlsShortLivedCache, AppConfig}
-import connectors.DataCacheConnector.cacheConnector
+import config.AppConfig
 import javax.inject.Inject
 import play.api.Logger
-import play.api.http.Status._
-import play.api.libs.json
-import play.api.libs.json._
+import play.api.libs.json.{JsValue, Json, Reads, Writes}
 import play.modules.reactivemongo.MongoDbConnection
 import uk.gov.hmrc.cache.TimeToLive
 import uk.gov.hmrc.cache.model.Cache
 import uk.gov.hmrc.cache.repository.CacheRepository
-import uk.gov.hmrc.crypto.json.{JsonDecryptor, JsonEncryptor}
 import uk.gov.hmrc.crypto.{ApplicationCrypto, CompositeSymmetricCrypto, Protected}
-import uk.gov.hmrc.http.cache.client.{CacheMap, ShortLivedCache}
-import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
-import uk.gov.hmrc.play.frontend.auth.AuthContext
-
-import scala.concurrent.ExecutionContext.Implicits.global
+import uk.gov.hmrc.crypto.json.{JsonDecryptor, JsonEncryptor}
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import scala.concurrent.Future
 
-trait CacheConnector {
-
-  def fetch[T](cacheId: String)(implicit authContext: AuthContext, hc: HeaderCarrier, formats: Format[T]): Future[Option[T]]
-
-  def save[T](cacheId: String, data: T)(implicit authContext: AuthContext, hc: HeaderCarrier, format: Format[T]): Future[CacheMap]
-
-  def fetchAll(implicit hc: HeaderCarrier, authContext: AuthContext): Future[Option[CacheMap]]
-
-  def remove(implicit hc: HeaderCarrier, ac: AuthContext): Future[HttpResponse]
-
-  def update[T](cacheId: String)(f: Option[T] => T)(implicit ac: AuthContext, hc: HeaderCarrier, fmt: Format[T]): Future[Option[T]]
-}
-
-class S4LCacheConnector @Inject()() extends CacheConnector {
-
-  lazy val shortLivedCache: ShortLivedCache = AmlsShortLivedCache
-
-  override def fetch[T](cacheId: String)
-                       (implicit authContext: AuthContext, hc: HeaderCarrier, formats: json.Format[T]): Future[Option[T]] =
-    shortLivedCache.fetchAndGetEntry[T](authContext.user.oid, cacheId)
-
-  override def save[T](cacheId: String, data: T)
-                      (implicit authContext: AuthContext, hc: HeaderCarrier, format: Format[T]): Future[CacheMap] =
-    shortLivedCache.cache(authContext.user.oid, cacheId, data)
-
-  override def fetchAll(implicit hc: HeaderCarrier, authContext: AuthContext): Future[Option[CacheMap]] =
-    shortLivedCache.fetch(authContext.user.oid)
-
-  override def remove(implicit hc: HeaderCarrier, ac: AuthContext): Future[HttpResponse] =
-    shortLivedCache.remove(ac.user.oid)
-
-  override def update[T](cacheId: String)(f: Option[T] => T)(implicit ac: AuthContext, hc: HeaderCarrier, fmt: Format[T]): Future[Option[T]] = {
-    fetch(cacheId) flatMap { t =>
-      val $t = f(t)
-      save(cacheId, $t) map { _ => Some($t) }
-    }
-  }
-}
-
-trait DataCacheConnector extends CacheConnector {
-  def cacheConnector: CacheConnector
-
-  override def fetch[T](cacheId: String)(implicit authContext: AuthContext, hc: HeaderCarrier, formats: Format[T]): Future[Option[T]] =
-    cacheConnector.fetch(cacheId)
-
-  override def save[T](cacheId: String, data: T)(implicit authContext: AuthContext, hc: HeaderCarrier, format: Format[T]): Future[CacheMap] =
-    cacheConnector.save(cacheId, data)
-
-  override def fetchAll(implicit hc: HeaderCarrier, authContext: AuthContext): Future[Option[CacheMap]] =
-    cacheConnector.fetchAll
-
-  override def remove(implicit hc: HeaderCarrier, ac: AuthContext): Future[HttpResponse] =
-    cacheConnector.remove
-
-  override def update[T](cacheId: String)(f: Option[T] => T)(implicit ac: AuthContext, hc: HeaderCarrier, fmt: Format[T]): Future[Option[T]] =
-    cacheConnector.update(cacheId)(f)
-}
-
-object DataCacheConnector extends DataCacheConnector {
-  def cacheConnector: CacheConnector = new S4LCacheConnector()
-}
-
-class AmlsMongoCacheWrapper @Inject()(mongoCache: AmlsMongoCache) extends CacheConnector {
-
-  def toCacheMap(c: Cache): CacheMap = {
-    c.data.fold(CacheMap(c.id.id, Map())) { json =>
-      ???
-    }
-  }
-
-  override def fetch[T](cacheId: String)(implicit authContext: AuthContext, hc: HeaderCarrier, formats: Format[T]): Future[Option[T]] = {
-    mongoCache.find(authContext.user.oid, cacheId)
-  }
-
-  override def save[T](cacheId: String, data: T)(implicit authContext: AuthContext, hc: HeaderCarrier, format: Format[T]): Future[CacheMap] = {
-    mongoCache.createOrUpdate(authContext.user.oid, data, cacheId) map { r =>
-      CacheMap(r.id.id, Map("data" -> r.data.get))
-    }
-  }
-
-  override def fetchAll(implicit hc: HeaderCarrier, authContext: AuthContext): Future[Option[CacheMap]] = {
-    mongoCache.cacheRepository.findById(authContext.user.oid)
-    ???
-  }
-
-  override def remove(implicit hc: HeaderCarrier, authContext: AuthContext): Future[HttpResponse] = {
-    mongoCache.removeById(authContext.user.oid) map {
-      case true => HttpResponse(OK)
-      case _ => HttpResponse(INTERNAL_SERVER_ERROR)
-    }
-  }
-
-  override def update[T](cacheId: String)(f: Option[T] => T)(implicit ac: AuthContext, hc: HeaderCarrier, fmt: Format[T]): Future[Option[T]] = ???
-}
-
-class AmlsMongoCache @Inject()(appConfig: AppConfig) extends MongoDbConnection with TimeToLive {
+class AmlsMongoCacheClient @Inject()(appConfig: AppConfig) extends MongoDbConnection with TimeToLive {
   implicit val compositeSymmetricCrypto: CompositeSymmetricCrypto = ApplicationCrypto.JsonCrypto
 
   private val expireAfter: Long = defaultExpireAfter
@@ -150,16 +48,16 @@ class AmlsMongoCache @Inject()(appConfig: AppConfig) extends MongoDbConnection w
     cacheRepository.createOrUpdate(id, key, jsonData).map(_.updateType.savedValue)
   }
 
-//  def createOrUpdateJson(id: String, json: JsValue, key: String = defaultKey): Future[Cache] = {
-//    val jsonData = if(appConfig.mongoEncryptionEnabled){
-//      val jsonEncryptor = new JsonEncryptor[JsValue]()
-//      Json.toJson(Protected(json))(jsonEncryptor)
-//    } else {
-//      json
-//    }
-//
-//    cacheRepository.createOrUpdate(id, key, jsonData).map(_.updateType.savedValue)
-//  }
+  //  def createOrUpdateJson(id: String, json: JsValue, key: String = defaultKey): Future[Cache] = {
+  //    val jsonData = if(appConfig.mongoEncryptionEnabled){
+  //      val jsonEncryptor = new JsonEncryptor[JsValue]()
+  //      Json.toJson(Protected(json))(jsonEncryptor)
+  //    } else {
+  //      json
+  //    }
+  //
+  //    cacheRepository.createOrUpdate(id, key, jsonData).map(_.updateType.savedValue)
+  //  }
 
   def createOrUpdateSeq[T](id: String, data: Seq[T], key: String = defaultKey)(implicit writes: Writes[T]): Future[Seq[T]] = {
     val jsonData = if(appConfig.mongoEncryptionEnabled){
