+++ test/connectors/cache/DataCacheConnectorSpec.scala
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package connectors
+package connectors.cache
 
-import config.AmlsShortLivedCache
 import org.mockito.Matchers.{eq => eqTo, _}
 import org.mockito.Mockito._
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.mock.MockitoSugar
 import org.scalatestplus.play.{OneAppPerSuite, PlaySpec}
-import play.api.libs.json.{Format, Json}
+import play.api.libs.json.Json
 import uk.gov.hmrc.http.cache.client.{CacheMap, ShortLivedCache}
+import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.play.frontend.auth.{AuthContext, LoggedInUser}
 
 import scala.concurrent.Future
-import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 
 class DataCacheConnectorSpec
   extends PlaySpec
@@ -55,7 +54,7 @@ class DataCacheConnectorSpec
 
   val emptyCache = CacheMap("", Map.empty)
 
-  object DataCacheConnector extends S4LCacheConnector {
+  object DataCacheConnector extends Save4LaterCacheConnector {
     override lazy val shortLivedCache: ShortLivedCache = mock[ShortLivedCache]
   }
